HOCS-4821: support hotfixing (#26)
Update GitHub actions so that on a PR it checks an additional
`skip-release` label. There is also a dry-run on the semver-tag-action
to ensure that the new tag is valid for that repository.

If `skip-release` is specified then no semver tag is generated. This
should be used for instances where changes don't impact either code or a
deployment i.e. README changes.